Data Pipeline V2: Update Audio to new DataModule (#1044)
