+++ flash/audio/classification/input_transform.py
@@ -11,12 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from dataclasses import dataclass
 from typing import Callable, Dict, Optional, Tuple
 
 import torch
 from torch import nn
 from torch.utils.data._utils.collate import default_collate
 
+from flash.core.data.input_transform import InputTransform
 from flash.core.data.io.input import DataKeys
 from flash.core.data.transforms import ApplyToKeys, merge_transforms
 from flash.core.utilities.imports import _TORCHAUDIO_AVAILABLE, _TORCHVISION_AVAILABLE
@@ -55,3 +57,28 @@ def train_default_transforms(
     if len(augs) > 0:
         return merge_transforms(default_transforms(spectrogram_size), {"per_sample_transform": nn.Sequential(*augs)})
     return default_transforms(spectrogram_size)
+
+
+@dataclass
+class AudioClassificationInputTransform(InputTransform):
+
+    spectrogram_size: Tuple[int, int] = (128, 128)
+    time_mask_param: Optional[int] = None
+    freq_mask_param: Optional[int] = None
+
+    def train_input_per_sample_transform(self) -> Callable:
+        transforms = []
+        if self.time_mask_param is not None:
+            transforms.append(TAudio.TimeMasking(time_mask_param=self.time_mask_param))
+
+        if self.freq_mask_param is not None:
+            transforms.append(TAudio.FrequencyMasking(freq_mask_param=self.freq_mask_param))
+
+        transforms += [T.ToTensor(), T.Resize(self.spectrogram_size)]
+        return T.Compose(transforms)
+
+    def input_per_sample_transform(self) -> Callable:
+        return T.Compose([T.ToTensor(), T.Resize(self.spectrogram_size)])
+
+    def target_per_sample_transform(self) -> Callable:
+        return torch.as_tensor
