+++ flash_examples/audio_classification.py
@@ -24,7 +24,8 @@ download_data("https://pl-flash-data.s3.amazonaws.com/urban8k_images.zip", "./da
 datamodule = AudioClassificationData.from_folders(
     train_folder="data/urban8k_images/train",
     val_folder="data/urban8k_images/val",
-    spectrogram_size=(64, 64),
+    transform_kwargs=dict(spectrogram_size=(64, 64)),
+    batch_size=4,
 )
 
 # 2. Build the model.
@@ -40,7 +41,8 @@ datamodule = AudioClassificationData.from_files(
         "data/urban8k_images/test/air_conditioner/13230-0-0-5.wav.jpg",
         "data/urban8k_images/test/children_playing/9223-2-0-15.wav.jpg",
         "data/urban8k_images/test/jackhammer/22883-7-10-0.wav.jpg",
-    ]
+    ],
+    batch_size=3,
 )
 predictions = trainer.predict(model, datamodule=datamodule)
 print(predictions)
